requirements.txt
appdirs==1.4.4
asgiref==3.3.4
auditlog3==1.0.1
black==20.8b1
click==7.1.2
Django==3.2
django-jazzmin==2.4.4
django-letsencrypt==4.0.0
feedparser==6.0.2
flake8==3.9.1
jsonfield==3.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
regex==2021.4.4
sgmllib3k==1.0.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
whitenoise==5.2.0
